Separate GetPropertiesWithNull
